:sparkles: Added a last updated at to the README file.

+++ build_readme/cli.py
@@ -9,10 +9,13 @@ README_PATH = ROOT_PATH / "README.md"
 
 
 @click.command
-def cli():
+@click.option("--force", "-f", is_flag=True)
+def cli(force: bool = False):
     """Build README.md"""
     with README_PATH.open("r") as file_obj:
         readme_content = file_obj.read()
+
+    org_readme_content = readme_content
 
     blog_posts = []
     remote_blog_posts = service.get_blog_posts()
@@ -34,5 +37,15 @@ def cli():
         readme_content, marker="MICROBLOG_POSTS", chunk="\n".join(microblog_posts)
     )
 
+    if org_readme_content == readme_content and force is False:
+        click.echo("No new changes to the README file.")
+        return
+
+    readme_content = service.replace_chunk(
+        readme_content, marker="LAST_UPDATED_AT", chunk=service.format_last_updated_at()
+    )
+
     with README_PATH.open("w") as file_obj:
         file_obj.write(readme_content)
+
+    click.echo("Updated the README file.")
